refactor(types): move common types to file
